build: Let's use bootleg
Because writing is such a great ability. It is thought that human beings developed language c. 35,000 BCE as evidenced by cave paintings from the period of the Cro-Magnon Man (c. 50,000-30,000 BCE) which appear to express concepts concerning daily life. These images suggest a language because, in some instances, they seem to tell a story (say, of a hunting expedition in which specific events occurred) rather than being simply pictures of animals and people. Written language, however, does not emerge until its invention in Sumer, southern Mesopotamia, c. 3500 -3000 BCE. This early writing was called cuneiform and consisted of making specific marks in wet clay with a reed implement. The writing system of the Egyptians was already in use before the rise of the Early Dynastic Period (c. 3150 BCE) and is thought to have developed from Mesopotamian cuneiform (though this theory is disputed) and came to be known as heiroglyphics.

tl;dr: It's all junk talk about writing. Let's set it this way because we're lazy to write gers and welp, it's too long tbh